made it float

+++ Engine/Rect.cpp
@@ -1,6 +1,6 @@
 #include "Rect.h"
 
-Rect::Rect(int x, int y, const int H, const int W, Color C):
+Rect::Rect(float x, float y, const float H, const float W, Color C):
 center(x,y),
 Height(H),
 Width(W),
@@ -13,34 +13,34 @@ Vec2 Rect::GetCenter() const
 	return center;
 }
 
-const int Rect::GetH() const
+const float Rect::GetH() const
 {
 	return Height;
 }
 
-const int Rect::GetW() const
+const float Rect::GetW() const
 {
 	return Width;
 }
 
 const Vec2 Rect::GetUR() const
 {
-	return this->center + Vec2(Width/2,-Height/2);
+	return this->center + Vec2(Width/2.0f,-Height/2.0f);
 }
 
 const Vec2 Rect::GetUL() const
 {
-	return this->center + Vec2(-Width / 2, -Height / 2);
+	return this->center + Vec2(-Width / 2.0f, -Height / 2.0f);
 }
 
 const Vec2 Rect::GetDR() const
 {
-	return this->center + Vec2(Width / 2, Height / 2);
+	return this->center + Vec2(Width / 2.0f, Height / 2.0f);
 }
 
 const Vec2 Rect::GetDL() const
 {
-	return this->center + Vec2(-Width / 2, Height / 2);
+	return this->center + Vec2(-Width / 2.0f, Height / 2.0f);
 }
 
 const Color Rect::GetC() const
